application scenario
User class is tested, and other are being tested. The application scenario is set as follows:

I want to create a loop within main to show a menu with the following options:
(1) Create a new document.
(2) Search for a document.
(3) Update document data.
(X) to exit the system.

I'll ask the user to select and enter a number from the following list numbers.