A GUI.py is created and being developed
Based on my revised specs and milestones, I'm working towards a "finished" application. I've creatd a GUI.py and my plan is to make the code in GUI.py  runs the Family-Doc-Sys directly. The main program will start the main window and the program keeps running until we close it.