The Spec 2 versions and more done in GUIDocumentHandler and DocumentManagementSystemGUI classes
The Spec 2 version has been uploaded and parts of the file have been completed. In the Spec 2 version, an explanation of the changes that have been made or are being worked on. The interfaces have been redesigned based on the correction comments. Parts of the GUIDocumentHandler class code have been completed, including the function to create a new document, and to modify an existing document.